requirements.txt
fastapi==0.115.2
starlette==0.40.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic>=2.9.0
pymongo==4.8.0
requests==2.32.3
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
